Support for btGhostObjects in the Bullet API.
Ghost objects are pure collision objects that have no effect on rigid
bodies, but can be used to hook into the collision detection of Bullet.

This way Blender can utilize the optimized collision detection in Bullet
as well as its contact caching, which is necessary for stable collision
response in many softbody-type algorithms that implement Coulomb
friction cones.