Initial commit for adaptive resolution.
Basic particle splitting method due to Feldman & Bonet 2007 IJNME (FB07)
implemented for classical SPH solver. In contrast to FB07 the implemented
method is for 3-dimensional domains and splits each particle into 9
daughters arranged in a cubic layout.

New members added to ParticleData struct to flag particle splitting and
store factors to scale particle mass and smoothing length appropriately.

New member added to ParticleSystem struct to keep track of total number
of particles split.

Solver routines have been modified to account for variable mass and
smoothing length. This basically amounts to scaling the mass and
smoothing length by the factors stored in ParticleData, and symmetrizing
the kernel used in the force callback in order to enforce momentum
conservation (See Monaghan 1992 Annu Rev Astron Astrophys). Note a
symmetrized kernel is not used in the density callback.

Splitting is currently toggled via a preprocessor directive and is hard
coded to occur if a particle is in a specified part of the domain on a
specified frame.