First attempt at particle coalescing.
Functions to coalesce particles using Vacondio et al's method have
been implemented.

Simple dam break test shows basic procedure works, particles are
split and then coalesced but the point cache does not work with the
addition of coalescing and the simulation result is not nice, though
this could be because of the particular simulation setup used and/or
the splitting/coalescing locations.