Added pressure as an output parameter of the SPH field sampler.
Pressure is now reported in addition to density. This required a
refactor of some of the SPH code. Two new callback functions were
added: one to convert density to pressure (the equation of state),
and another to initialise the input parameters that are needed by
the former. This also required a couple of parameters to be added
to the SPHData struct, and a new struct for storing the outputs.