BVH tree now used for coalescing nearest neighbour search, various bugs also fixed and changes made to sphmassfac's used for splitting.
- Particle coalescing now uses BVH tree to find nearest
neighbours for coalescing. This allows particles to merge
with any other split particle in the domain and this greatly
improves simulation quality.

- Mass factors used for splitting are now 0.2 and 0.1. This
is convenient as coalescing is now much more likely to
produce particles with sphmassfac = 1.0.

- Maximum sphmassfac through coalescing is also now limited
to 1.1.

- If a particles sphmassfac after merging is >= 1.0,
sphmassfac and sphalpha are set to 1.0 so that the particle
is effectively reset to it's original size and it is now
allowed to be split again. This is important because if
particles about the size of the original particles are
created through coalescing and are not allowed to be
re-split again, they will eventually fill the splitting
region and the adaptive resolution will cease to have an
effect.

- The split through wall test function now checks if there
are any colliders before attempting to check for wall
intersects. This is necessary to prevent Blender from
crashing if there are no colliders.

- Particles are set as 'DEAD' inside the coalescing function
now to avoid merging particles with particles that have
already been set to die in the next frame via the dietime
variable. Failure to do this was causing all split particles
to be set as 'DEAD' through the coalescing function.