Removed printf statements.
Printf statements used to verify particle splitting is working as
intended have been removed. These statements were used to verify
that the code applies coalescing to particles with appropriate
sphmassfac only, correctly coalesces particles such that no particles
with sphmassfac > 1.05 are created, only resets coalesced particles
with sphmassfac = 1.0 to a state in which they can be split again
and never re-initializes the adaptive resolution variables
(sphmassfac, sphalpha and split) after frame 0.

The simulations also verified that the code is coalescing split
particles in to particles with sphmassfac = 1.0 very quickly after
they leave the adaptive resolution region. This is crucial to the
effectiveness of the adaptive resolution method.