Basic refiners implemented.
Particles are now split if they are within a refiner region
defined in the UI.

For mesh objects, refinement regions can be defined as locations
within a user defined radius of either mesh faces or edges. Mesh
vertices will be added at a later date.

For empty objects, refinement regions can be defined as locations
within a user defined radius of the objects location.

Currently particles are split (merged) if they are (are not) in
one of these user defined splitting regions. In the future,
particle splitting will be determined by additional requirements
for min/max particle mass in a refinement region.