Temporary bug fix for threading issue.
The refiner object derivedFinal member can be de-allocated by
a different function running on a separate thread when
refiners are being initialized. This leads to the refiner not
being added to the refiners list and it having no effect on
the particles. This issue is most obvious in small scale test
simulations and was not obvious in larger simulations so it
may not be an issue in practice. As best I can tell, the same
issue should affect the existing effectors code.

A temporary solution has been put in for the purpose of testing
where a while loop is used to pause the execution of the effected
parts of the refiner initialization until the refiners derivedFinal
is non-null.