Bug fix.
Dynamic particle birth and die times needed for particle re-use
caused particles to erroneously appear in the wrong positions
when reading from point cache. The stored location of an
incorrectly positioned particle was correct when checked in a
debugger or via the python API but it instantaneously appeared
in the UI at another location before being corrected on the next
frame.

The cause of the issue appears to be that the particles time
variables (time, dietime, lifetime) were not updated from the
point cache before they were used to set pa->state.time. The
appropriate value to set for pa->state.time is dependent on a
comparison of pa->time and pa->dietime to the current frame
number being read from point cache. This comparison was done
before the time values were read from point cache resulting in
an incorrect value for pa->state.time.

The fix for this issue was to re-order these steps so that the
time variables are updated before pa->state.time is set.