Custom data layer implemented for refiner radius.
Refiners using distance from mesh faces with the radius parameter
now store the refiner radius as a custom data layer in the refiner
object. This allows the radius to be set separately for each face
which allows the search for the shortest distance from a particle
to a face to take advantage of acceleration structures.

For example, rather than having to iterate over 100 refiner objects
with different refiner radii to check which is closest, the 100
objects can be merged into a single object with radii set separately
for the appropriate faces. When finding the nearest face then, the
code only has to query a single refiner object where a BVHTree made
from the objects faces is used to more efficiently search for the
nearest face. Once the nearest face is found the appropriate radius
value is retrieved from the corresponding location in the custom
data layer.